feat: move api to the store
